Sort README props table alphabetically
Thank you so much for this library and its ongoing maintenance and development - I love it and depend on it in many ways. 

I wish to contribute a documentation PR whose need is prob imaginary on my part, but is arguably more in-line with typical conventions or assumptions ( at least as far as random tables of text on the internet are concerned ). 

I could not discern a pattern or sort-order by which this table was previously ordered, however I found myself frequently wishing it were so alphabetically. I ask that you consider this PR as I frequently would look for props whose exact spelling I did not know or remember, yet at least had a good guess for the first few letters, or let me know if anything else is requested on my part.

If may have missed any preferred sort-order, I apologize!!

Also, I used the github "edit file" feature - I hope I did not mess up any templates or PR rules. Please let me know and I'll remake this ASAP.

Thanks again for all the hard thankless work.